Fetch 5xx and total error rates from graphite
It's possible I can extract this from one API call, but I'm not
sure how. This makes two calls and then tries to reconcile
the two CSVs.

Currently it fetches all data for the preceding month, for each
app.

This is a first step towards automating calculation of overall
error rates per application.

We're interested in HTTP errors as a metric because it's
comparable across most of our apps, unlike sentry errors
or zendesk tickets, which vary depending on reporting behaviour.