Structure product in terms of utilities, steps, and pipelines
This is inspired by the verify IDP data pipeline.

Pipeline utilities should be reusable across steps (i.e. not coupled
to business logic)

In verify's pipeline, steps can be reused in multiple sequences, but
I've chosen not implement this until there are multiple pipelines.

A step is not coupled to the workflow library. The plan is to adopt
Luigi for this.