Add a step to upload to google drive
This is an easy way of sharing the reports which will be
generated by this pipeline, and using the data in google data studio.

Ideally this would use google team drive, which the library supports,
but this will need more work to set up so I'm going to come back to
this. For now, the service user can share spreadsheets with a mailing
list.

I've extracted out snapshotting and directory management so we
can load the ouputs of earlier steps.