Change snapshot behaviour
Snapshots are supposed to be saved as soon as possible, so you can debug
what happened whether the pipeline ran or failed.

But the target of a luigi task should only be created on success,
so that running the pipeline is idempotent.

To reconcile these two things, make the target a separate file.